+++ controllers/index.ts
@@ -1,0 +1,4 @@
+export * from "./userController";
+export * from "./authController";
+export * from "./menuItemController";
+export * from "./restaurantController";
